+++ raiden/ui/web/src/app/components/channel-table/channel-table.component.spec.ts
@@ -26,6 +26,7 @@ import { AddressInputComponent } from '../address-input/address-input.component'
 import { EventListComponent } from '../event-list/event-list.component';
 import { OpenDialogComponent } from '../open-dialog/open-dialog.component';
 import { TokenInputComponent } from '../token-input/token-input.component';
+import { TokenNetworkSelectorComponent } from '../token-network-selector/token-network-selector.component';
 
 import { ChannelTableComponent } from './channel-table.component';
 import Spy = jasmine.Spy;
@@ -72,7 +73,8 @@ describe('ChannelTableComponent', () => {
                 KeysPipe,
                 SubsetPipe,
                 DecimalPipe,
-                CdkDetailRowDirective
+                CdkDetailRowDirective,
+                TokenNetworkSelectorComponent
             ],
             providers: [
                 SharedService,