+++ raiden/ui/web/src/app/components/payment-dialog/payment-dialog.component.spec.ts
@@ -10,6 +10,7 @@ import { SharedService } from '../../services/shared.service';
 import { AddressInputComponent } from '../address-input/address-input.component';
 import { MockConfig } from '../channel-table/channel-table.component.spec';
 import { TokenInputComponent } from '../token-input/token-input.component';
+import { TokenNetworkSelectorComponent } from '../token-network-selector/token-network-selector.component';
 import { PaymentDialogComponent, PaymentDialogPayload } from './payment-dialog.component';
 
 describe('PaymentDialogComponent', () => {
@@ -28,7 +29,8 @@ describe('PaymentDialogComponent', () => {
                 PaymentDialogComponent,
                 TokenPipe,
                 TokenInputComponent,
-                AddressInputComponent
+                AddressInputComponent,
+                TokenNetworkSelectorComponent
             ],
             providers: [
                 {